Extracted RegexGroupParser from RegexArrayShapeMatcher
