+++ conf/config.neon
@@ -1501,6 +1501,9 @@ services:
 		class: PHPStan\Type\Php\RegexArrayShapeMatcher
 
 	-
+		class: PHPStan\Type\Php\RegexGroupParser
+
+	-
 		class: PHPStan\Type\Php\RegexExpressionHelper
 
 	-
